# [misc] Optimize speculative decoding