This PR introduces optimizations to the speculative decoding process by reducing memory copy overhead, resulting in improved performance.

After profiling, I found that cuda memcpy occurred abnormally, so I improved the code. (Refer to the profiling image below)
I experimented with one a100 gpu, and the code I used for the experiment is as follows.

I improved the `_filter_model_output` function and `_merge_outputs` function, and confirmed that the test code below improved by about 3%.

| Applied Function | E2E Latency |
| --- | --- |
| baseline | 0.462s |
| `_filter_model_output` | 0.457s |
| `_merge_outputs` | 0.459s |
| `_filter_model_output` + `_merge_outputs` | 0.449s |

<details>
<summary>Test code using the OPT model</summary>

```python
import time
import nvtx
from typing import List
from vllm import LLM, SamplingParams

def time_generation(llm: LLM, prompts: List[str],
                    sampling_params: SamplingParams):
    llm.generate(prompts, sampling_params)
    llm.generate(prompts, sampling_params)
    times = []
    with nvtx.annotate("generate", color="blue"):
        for _ in range(10):
            start = time.time()
            outputs = llm.generate(prompts, sampling_params)
            end = time.time()
            times.append(end-start)
    print(sum(times) / len(times))

    for output in outputs:
        generated_text = output.outputs[0].text
        print(f"text: {generated_text!r}")


if __name__ == "__main__":
    sampling_params = SamplingParams(temperature=0.0, max_tokens=100)

    prompts = [
        "The president of the United States is",
    ]

    llm = LLM(
        model="facebook/opt-6.7b",
        tensor_parallel_size=1,
        speculative_model="facebook/opt-125m",
        num_speculative_tokens=5,
        use_v2_block_manager=True,
    )

    time_generation(llm, prompts, sampling_params)
```
</details>


**Profiling results**
- Before code improvement
        <img width="2135" alt="2024-08-27_00-08-21" src="https://github.com/user-attachments/assets/b66f43f3-b0ca-4779-af2f-f80f6912ae94">

- After code improvement
        <img width="2095" alt="2024-08-27_00-08-28" src="https://github.com/user-attachments/assets/ea494f45-b399-49df-a943-1b935dac1f58">