@jacob-crux 2024-08-27
> Thanks @jacob-crux! It looks like this mostly comes down to indexing using an on-device tensor rather than a list. This can also be used with the existing `tensor[index_tensor]` syntax ... are the `index_` tensor methods much faster?

Hello!
First, I confirmed that when slicing with `output_indices_to_retain`, which is a `List[int]` in the `_filter_model_output` function, `cudaMemcpy` is used to copy and sync the index every time.

When comparing the speed difference between the list slice kernel and `torch.index_select` with the test code, I found that there was a speed difference of about 40-50% when comparing only the kernel. However, since the proportion of the existing kernel is small, it is difficult to get a big feeling.

Below are the profiling results. The first result is from the existing code, the second result is when only the index is assigned to a CUDA tensor, and the last result is from the code using `torch.index_select` in the PR.
<img width="2132" alt="2024-08-27_15-07-53" src="https://github.com/user-attachments/assets/e81ad079-144d-4cd4-851a-c2c3a1e4aae0">