@njhill 2024-08-27
Thanks @jacob-crux, I wasn't questioning that using repeated list-based slicing is slower. I was suggesting that the same syntax can be  used to slice using a tensor. Then the actual code modifications needed would be minimal.

i.e.
```python
index_list = [1, 3, 5]
index_tensor = torch.tensor(index_list, device=device)
sliced = a_tensor[index_tensor]
sliced2 = another_tensor[index_tensor]
```