@jacob-crux 2024-08-28
> Thanks @jacob-crux, I wasn't questioning that using repeated list-based slicing is slower. I was suggesting that the same syntax can be used to slice using a tensor. Then the actual code modifications needed would be minimal.
> 
> i.e.
> 
> ```python
> index_list = [1, 3, 5]
> index_tensor = torch.tensor(index_list, device=device)
> sliced = a_tensor[index_tensor]
> sliced2 = another_tensor[index_tensor]
> ```

@njhill Thank you for your review.
[The comment](https://github.com/vllm-project/vllm/pull/7875#issuecomment-2311657316) was an experiment with the code you mentioned.

Here is a simple code example of the 3 profiled contents in the picture in the comment.

```python
# profile 1
index_list = [1, 3, 5]
sliced = a_tensor[index_tensor]

# profile 2
index_list = [1, 3, 5]
index_tensor = torch.tensor(index_list, device=device)
sliced = a_tensor[index_tensor]

# profile 3
index_list = [1, 3, 5]
index_tensor = torch.tensor(index_list, device=device)
sliced = a_tensor.index_select(0, index_tensor)
```
The difference mentioned here is
- The difference between profile 1 and profile 2 is that memcpy occurs frequently.
- The difference between profile 1&profile 2 and profile 3 is the difference in the GPU kernel called when slicing, and as in the middle section of the picture, the kernel shows a speed difference of about 40~50%.

Although the code of profile 2 may look a little simpler, the kernel of profile 3 has the effect of improving performance on the GPU.